requirements.txt
Django~=5.1.5
djangorestframework~=3.15.2
django-filter~=24.3
google~=3.0.0
python-dotenv~=1.0.1
google-auth==2.38.0
django-cors-headers==4.6.0
drf-spectacular==0.28.0
drf-yasg==1.21.8
pillow==11.1.0
psycopg2==2.9.10
requests==2.32.3
PyJWT==2.10.1
urllib3==2.3.0
uvicorn==0.34.0

